+++ server/utils/auth.js
@@ -6,9 +6,9 @@ const expiration = '2h';
 
 module.exports = {
   // function for our authenticated routes
-  authMiddleware: function (req, res, next) {
+  authMiddleware: function ({req}) {
     // allows token to be sent via  req.query or headers
-    let token = req.query.token || req.headers.authorization;
+    let token = req.query.token || req.headers.authorization || req.body.token;
 
     // ["Bearer", "<tokenvalue>"]
     if (req.headers.authorization) {
@@ -16,7 +16,8 @@ module.exports = {
     }
 
     if (!token) {
-      return res.status(400).json({ message: 'You have no token!' });
+      //return res.status(400).json({ message: 'You have no token!' });
+      return req
     }
 
     // verify token and get user data out of it
@@ -29,10 +30,10 @@ module.exports = {
     }
 
     // send to next endpoint
-    next();
+    return req;
   },
-  signToken: function ({ username, email, _id }) {
-    const payload = { username, email, _id };
+  signToken: function ({ username, email, _id, age }) {
+    const payload = { username, email, _id, age };
 
     return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
   },
